match cpu/gpu affine

+++ src/math.h
@@ -14,7 +14,7 @@ void cosWindow(double* cos, const int w, const int h);
 void guassian2d(double* guass, const int w, const int h);
 void dft2d(const int M, const int N, double* f, double* F);
 void idft2d(const int M, const int N, double* F, double* f);
-void affine(double* src, int sw, int sh, double* dst, int dw, int dh, double m[2][3]);
+void affine(uint8_t* src, int sw, int sh, uint8_t* dst, int dw, int dh, double m[2][3]);
 void preproc(double* f, double* cos, double* dst, int w, int h);
 
 #ifdef USE_OPENCV
